pagi-intelligence-bridge/pyproject.toml
[tool.poetry]
name = "pagi-intelligence-bridge"
version = "0.1.0"
description = "Python RLM Intelligence for Phoenix AGI"
authors = ["Sovereign Architect"]

[tool.poetry.dependencies]
python = "^3.10"
# FastAPI >=0.100 supports Pydantic v2; keep aligned with codebase (pydantic>=2).
fastapi = "^0.117"
uvicorn = "^0.21"
litellm = "^0.1"
pydantic = "^2.0"
sentence-transformers = "^2.2"
grpcio = "^1.60"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
grpcio-tools = "^1.60"
# Required by starlette.testclient / fastapi.testclient.
httpx = "^0.28"
